chore: move chain families to separate chain.go file
