+++ tools/operations-gen/generate/generate.go
@@ -5,23 +5,15 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
-	"strings"
 	"text/template"
 
 	"gopkg.in/yaml.v3"
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/tools/operations-gen/internal/core"
-	"github.com/smartcontractkit/chainlink-deployments-framework/tools/operations-gen/internal/families/evm"
 )
 
 //go:embed templates
 var templatesFS embed.FS
-
-// chainFamilies is the single registration point for all supported chain families.
-var chainFamilies = map[string]core.ChainFamilyHandler{
-	"evm": evm.Handler{},
-}
 
 // Config holds the top-level generator configuration.
 // Input/Output/Contracts are raw YAML nodes so chain-family handlers own their
@@ -87,15 +79,4 @@ func LoadTemplate(chainFamily string) (*template.Template, error) {
 	}
 
 	return template.New("operations").Parse(string(content))
-}
-
-// supportedFamilies returns a sorted, comma-separated list of supported chain families.
-func supportedFamilies() string {
-	families := make([]string, 0, len(chainFamilies))
-	for k := range chainFamilies {
-		families = append(families, k)
-	}
-	sort.Strings(families)
-
-	return strings.Join(families, ", ")
 }
